+++ primefaces/src/main/java/org/primefaces/component/menu/MenuBase.java
@@ -45,7 +45,8 @@ public abstract class MenuBase extends AbstractMenu implements Widget, OverlayMe
         toggleable,
         statefulGlobal,
         collision,
-        maxHeight
+        maxHeight,
+        appendTo
     }
 
     public MenuBase() {
@@ -164,5 +165,13 @@ public abstract class MenuBase extends AbstractMenu implements Widget, OverlayMe
 
     public void setStatefulGlobal(boolean statefulGlobal) {
         getStateHelper().put(PropertyKeys.statefulGlobal, statefulGlobal);
+    }
+
+    public String getAppendTo() {
+        return (String) getStateHelper().eval(PropertyKeys.appendTo, "@(body)");
+    }
+
+    public void setAppendTo(String appendTo) {
+        getStateHelper().put(PropertyKeys.appendTo, appendTo);
     }
 }