+++ primefaces/src/main/resources/META-INF/resources/primefaces/menu/menu.base.js
@@ -59,7 +59,9 @@ PrimeFaces.widget.Menu = PrimeFaces.widget.BaseWidget.extend({
         //mark trigger and descandants of trigger as a trigger for a primefaces overlay
         this.trigger.data('primefaces-overlay-target', true).find('*').data('primefaces-overlay-target', true);
 
-        this.cfg.appendTo = '@(body)';
+        if (!this.cfg.appendTo) {
+            this.cfg.appendTo = '@(body)';
+        }
         PrimeFaces.utils.registerDynamicOverlay(this, this.jq, this.id);
         this.transition = PrimeFaces.utils.registerCSSTransition(this.jq, 'ui-connected-overlay');